requirements.txt
selenium
notebook
pandas
yfinance
scikit-learn
praw
transformers
torch
tqdm
python-dotenv
plotly